Moved around variables and states and changed functions
Fixing the issue of having my simulation state being drawn overtop of all states, moved the simulation(); function around so it wasn't constantly being redrawn at 30fps. Created an array and for loops for the stars to be drawn at a random size between two different values, also not constantly being drawn. Changed from mousePressed(); to click - part of the issue of the click not working when referring to the stars class.